hotels, flights, in global state
